requirements.txt
requests==2.32.3
pytest==8.3.1
allure-pytest==2.13.5
allure-python-commons==2.13.5